bug(stdlib): use size_t for the size param in malloc

+++ stdlib/memory.c
@@ -16,7 +16,7 @@ void memset(uint8_t *destination, uint8_t value, uint32_t amount) {
 // remember that the kernel starts at 0x1000 as defined in the Makefile
 uint32_t free_memory_address = 0x10000;
 
-uint32_t malloc(uint32_t size, int align, uint32_t *physical_address) { // the implementation is just a pointer to some free memory which keeps growing
+uint32_t malloc(size_t size, int align, uint32_t *physical_address) { // the implementation is just a pointer to some free memory which keeps growing
     // pages are aligned to 4K, or 0x1000
     if (align == 1 && (free_memory_address & 0xFFFFF000)) {
         free_memory_address &= 0xFFFFF000;
